+++ src/werkzeug/urls.py
@@ -59,7 +59,7 @@ class BaseURL(_URLTuple):
     """Superclass of :py:class:`URL` and :py:class:`BytesURL`.
 
     .. deprecated:: 2.3
-        Will be removed in Werkzeug 2.4. Use the ``urllib.parse`` library instead.
+        Will be removed in Werkzeug 3.0. Use the ``urllib.parse`` library instead.
     """
 
     __slots__ = ()
@@ -71,7 +71,7 @@ class BaseURL(_URLTuple):
     def __new__(cls, *args: t.Any, **kwargs: t.Any) -> BaseURL:
         warnings.warn(
             f"'werkzeug.urls.{cls.__name__}' is deprecated and will be removed in"
-            " Werkzeug 2.4. Use the 'urllib.parse' library instead.",
+            " Werkzeug 3.0. Use the 'urllib.parse' library instead.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -358,7 +358,7 @@ class URL(BaseURL):
     URL.
 
     .. deprecated:: 2.3
-        Will be removed in Werkzeug 2.4. Use the ``urllib.parse`` library instead.
+        Will be removed in Werkzeug 3.0. Use the ``urllib.parse`` library instead.
     """
 
     __slots__ = ()
@@ -371,22 +371,20 @@ class URL(BaseURL):
         """Encodes the URL to a tuple made out of bytes.  The charset is
         only being used for the path, query and fragment.
         """
-        with warnings.catch_warnings():
-            warnings.filterwarnings("ignore", "'werkzeug", DeprecationWarning)
-            return BytesURL(
-                self.scheme.encode("ascii"),
-                self.encode_netloc(),
-                self.path.encode(charset, errors),
-                self.query.encode(charset, errors),
-                self.fragment.encode(charset, errors),
-            )
+        return BytesURL(
+            self.scheme.encode("ascii"),
+            self.encode_netloc(),
+            self.path.encode(charset, errors),
+            self.query.encode(charset, errors),
+            self.fragment.encode(charset, errors),
+        )
 
 
 class BytesURL(BaseURL):
     """Represents a parsed URL in bytes.
 
     .. deprecated:: 2.3
-        Will be removed in Werkzeug 2.4. Use the ``urllib.parse`` library instead.
+        Will be removed in Werkzeug 3.0. Use the ``urllib.parse`` library instead.
     """
 
     __slots__ = ()
@@ -406,15 +404,13 @@ class BytesURL(BaseURL):
         """Decodes the URL to a tuple made out of strings.  The charset is
         only being used for the path, query and fragment.
         """
-        with warnings.catch_warnings():
-            warnings.filterwarnings("ignore", "'werkzeug", DeprecationWarning)
-            return URL(
-                self.scheme.decode("ascii"),  # type: ignore
-                self.decode_netloc(),
-                self.path.decode(charset, errors),  # type: ignore
-                self.query.decode(charset, errors),  # type: ignore
-                self.fragment.decode(charset, errors),  # type: ignore
-            )
+        return URL(
+            self.scheme.decode("ascii"),  # type: ignore
+            self.decode_netloc(),
+            self.path.decode(charset, errors),  # type: ignore
+            self.query.decode(charset, errors),  # type: ignore
+            self.fragment.decode(charset, errors),  # type: ignore
+        )
 
 
 _unquote_maps: dict[frozenset[int], dict[bytes, int]] = {frozenset(): _hextobyte}
@@ -504,10 +500,10 @@ def url_parse(
                             from the URL.
 
     .. deprecated:: 2.3
-        Will be removed in Werkzeug 2.4. Use ``urllib.parse.urlsplit`` instead.
+        Will be removed in Werkzeug 3.0. Use ``urllib.parse.urlsplit`` instead.
     """
     warnings.warn(
-        "'werkzeug.urls.url_parse' is deprecated and will be removed in Werkzeug 2.4."
+        "'werkzeug.urls.url_parse' is deprecated and will be removed in Werkzeug 3.0."
         " Use 'urllib.parse.urlsplit' instead.",
         DeprecationWarning,
         stacklevel=2,
@@ -546,9 +542,7 @@ def url_parse(
 
     result_type = URL if is_text_based else BytesURL
 
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore", "'werkzeug", DeprecationWarning)
-        return result_type(scheme, netloc, url, query, fragment)
+    return result_type(scheme, netloc, url, query, fragment)
 
 
 def _make_fast_url_quote(
@@ -605,13 +599,13 @@ def url_quote(
     :param unsafe: an optional sequence of unsafe characters.
 
     .. deprecated:: 2.3
-        Will be removed in Werkzeug 2.4. Use ``urllib.parse.quote`` instead.
+        Will be removed in Werkzeug 3.0. Use ``urllib.parse.quote`` instead.
 
     .. versionadded:: 0.9.2
        The `unsafe` parameter was added.
     """
     warnings.warn(
-        "'werkzeug.urls.url_quote' is deprecated and will be removed in Werkzeug 2.4."
+        "'werkzeug.urls.url_quote' is deprecated and will be removed in Werkzeug 3.0."
         " Use 'urllib.parse.quote' instead.",
         DeprecationWarning,
         stacklevel=2,
@@ -646,7 +640,7 @@ def url_quote_plus(
     :param safe: An optional sequence of safe characters.
 
     .. deprecated:: 2.3
-        Will be removed in Werkzeug 2.4. Use ``urllib.parse.quote_plus`` instead.
+        Will be removed in Werkzeug 3.0. Use ``urllib.parse.quote_plus`` instead.
     """
     warnings.warn(
         "'werkzeug.urls.url_quote_plus' is deprecated and will be removed in Werkzeug"
@@ -655,9 +649,7 @@ def url_quote_plus(
         stacklevel=2,
     )
 
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore", "'werkzeug", DeprecationWarning)
-        return url_quote(string, charset, errors, safe + " ", "+").replace(" ", "+")
+    return url_quote(string, charset, errors, safe + " ", "+").replace(" ", "+")
 
 
 def url_unparse(components: tuple[str, str, str, str, str]) -> str:
@@ -668,10 +660,10 @@ def url_unparse(components: tuple[str, str, str, str, str]) -> str:
                        into a URL string.
 
     .. deprecated:: 2.3
-        Will be removed in Werkzeug 2.4. Use ``urllib.parse.urlunsplit`` instead.
+        Will be removed in Werkzeug 3.0. Use ``urllib.parse.urlunsplit`` instead.
     """
     warnings.warn(
-        "'werkzeug.urls.url_unparse' is deprecated and will be removed in Werkzeug 2.4."
+        "'werkzeug.urls.url_unparse' is deprecated and will be removed in Werkzeug 3.0."
         " Use 'urllib.parse.urlunsplit' instead.",
         DeprecationWarning,
         stacklevel=2,
@@ -716,10 +708,10 @@ def url_unquote(
     :param errors: the error handling for the charset decoding.
 
     .. deprecated:: 2.3
-        Will be removed in Werkzeug 2.4. Use ``urllib.parse.unquote`` instead.
+        Will be removed in Werkzeug 3.0. Use ``urllib.parse.unquote`` instead.
     """
     warnings.warn(
-        "'werkzeug.urls.url_unquote' is deprecated and will be removed in Werkzeug 2.4."
+        "'werkzeug.urls.url_unquote' is deprecated and will be removed in Werkzeug 3.0."
         " Use 'urllib.parse.unquote' instead.",
         DeprecationWarning,
         stacklevel=2,
@@ -745,7 +737,7 @@ def url_unquote_plus(
     :param errors: The error handling for the `charset` decoding.
 
     .. deprecated:: 2.3
-        Will be removed in Werkzeug 2.4. Use ``urllib.parse.unquote_plus`` instead.
+        Will be removed in Werkzeug 3.0. Use ``urllib.parse.unquote_plus`` instead.
     """
     warnings.warn(
         "'werkzeug.urls.url_unquote_plus' is deprecated and will be removed in Werkzeug"
@@ -759,9 +751,7 @@ def url_unquote_plus(
     else:
         s = s.replace(b"+", b" ")
 
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore", "'werkzeug", DeprecationWarning)
-        return url_unquote(s, charset, errors)
+    return url_unquote(s, charset, errors)
 
 
 def url_fix(s: str, charset: str = "utf-8") -> str:
@@ -778,10 +768,10 @@ def url_fix(s: str, charset: str = "utf-8") -> str:
         as a string.
 
     .. deprecated:: 2.3
-        Will be removed in Werkzeug 2.4.
+        Will be removed in Werkzeug 3.0.
     """
     warnings.warn(
-        "'werkzeug.urls.url_fix' is deprecated and will be removed in Werkzeug 2.4.",
+        "'werkzeug.urls.url_fix' is deprecated and will be removed in Werkzeug 3.0.",
         DeprecationWarning,
         stacklevel=2,
     )
@@ -795,13 +785,11 @@ def url_fix(s: str, charset: str = "utf-8") -> str:
     if s.startswith("file://") and s[7:8].isalpha() and s[8:10] in (":/", "|/"):
         s = f"file:///{s[7:]}"
 
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore", "'werkzeug", DeprecationWarning)
-        url = url_parse(s)
-        path = url_quote(url.path, charset, safe="/%+$!*'(),")
-        qs = url_quote_plus(url.query, charset, safe=":&%=+$!*'(),")
-        anchor = url_quote_plus(url.fragment, charset, safe=":&%=+$!*'(),")
-        return url_unparse((url.scheme, url.encode_netloc(), path, qs, anchor))
+    url = url_parse(s)
+    path = url_quote(url.path, charset, safe="/%+$!*'(),")
+    qs = url_quote_plus(url.query, charset, safe=":&%=+$!*'(),")
+    anchor = url_quote_plus(url.fragment, charset, safe=":&%=+$!*'(),")
+    return url_unparse((url.scheme, url.encode_netloc(), path, qs, anchor))
 
 
 def _codec_error_url_quote(e: UnicodeError) -> tuple[str, int]:
@@ -868,7 +856,7 @@ def uri_to_iri(
 
     .. versionchanged:: 2.3
         Passing a tuple or bytes, and the ``charset`` and ``errors`` parameters, are
-        deprecated and will be removed in Werkzeug 2.4.
+        deprecated and will be removed in Werkzeug 3.0.
 
     .. versionchanged:: 2.3
         Which characters remain quoted is specific to each part of the URL.
@@ -882,7 +870,7 @@ def uri_to_iri(
     """
     if isinstance(uri, tuple):
         warnings.warn(
-            "Passing a tuple is deprecated and will not be supported in Werkzeug 2.4.",
+            "Passing a tuple is deprecated and will not be supported in Werkzeug 3.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -890,7 +878,7 @@ def uri_to_iri(
 
     if isinstance(uri, bytes):
         warnings.warn(
-            "Passing bytes is deprecated and will not be supported in Werkzeug 2.4.",
+            "Passing bytes is deprecated and will not be supported in Werkzeug 3.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -899,7 +887,7 @@ def uri_to_iri(
     if charset is not None:
         warnings.warn(
             "The 'charset' parameter is deprecated and will be removed"
-            " in Werkzeug 2.4.",
+            " in Werkzeug 3.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -908,7 +896,7 @@ def uri_to_iri(
 
     if errors is not None:
         warnings.warn(
-            "The 'errors' parameter is deprecated and will be removed in Werkzeug 2.4.",
+            "The 'errors' parameter is deprecated and will be removed in Werkzeug 3.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -960,7 +948,7 @@ def iri_to_uri(
 
     .. versionchanged:: 2.3
         Passing a tuple or bytes, and the ``charset`` and ``errors`` parameters, are
-        deprecated and will be removed in Werkzeug 2.4.
+        deprecated and will be removed in Werkzeug 3.0.
 
     .. versionchanged:: 2.3
         Which characters remain unquoted is specific to each part of the URL.
@@ -980,7 +968,7 @@ def iri_to_uri(
     """
     if isinstance(iri, tuple):
         warnings.warn(
-            "Passing a tuple is deprecated and will not be supported in Werkzeug 2.4.",
+            "Passing a tuple is deprecated and will not be supported in Werkzeug 3.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -988,7 +976,7 @@ def iri_to_uri(
 
     if isinstance(iri, bytes):
         warnings.warn(
-            "Passing bytes is deprecated and will not be supported in Werkzeug 2.4.",
+            "Passing bytes is deprecated and will not be supported in Werkzeug 3.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -997,7 +985,7 @@ def iri_to_uri(
     if charset is not None:
         warnings.warn(
             "The 'charset' parameter is deprecated and will be removed"
-            " in Werkzeug 2.4.",
+            " in Werkzeug 3.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -1006,7 +994,7 @@ def iri_to_uri(
 
     if errors is not None:
         warnings.warn(
-            "The 'errors' parameter is deprecated and will be removed in Werkzeug 2.4.",
+            "The 'errors' parameter is deprecated and will be removed in Werkzeug 3.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -1016,7 +1004,7 @@ def iri_to_uri(
     if safe_conversion is not None:
         warnings.warn(
             "The 'safe_conversion' parameter is deprecated and will be removed in"
-            " Werkzeug 2.4.",
+            " Werkzeug 3.0.",
             DeprecationWarning,
             stacklevel=2,
         )
@@ -1065,6 +1053,27 @@ def iri_to_uri(
     return urlunsplit((parts.scheme, netloc, path, query, fragment))
 
 
+def _invalid_iri_to_uri(iri: str) -> str:
+    """The URL scheme ``itms-services://`` must contain the ``//`` even though it does
+    not have a host component. There may be other invalid schemes as well. Currently,
+    responses will always call ``iri_to_uri`` on the redirect ``Location`` header, which
+    removes the ``//``. For now, if the IRI only contains ASCII and does not contain
+    spaces, pass it on as-is. In Werkzeug 3.0, this should become a
+    ``response.process_location`` flag.
+
+    :meta private:
+    """
+    try:
+        iri.encode("ascii")
+    except UnicodeError:
+        pass
+    else:
+        if len(iri.split(None, 1)) == 1:
+            return iri
+
+    return iri_to_uri(iri)
+
+
 def url_decode(
     s: t.AnyStr,
     charset: str = "utf-8",
@@ -1084,7 +1093,7 @@ def url_decode(
     :param cls: Container to hold result instead of :class:`MultiDict`.
 
     .. deprecated:: 2.3
-        Will be removed in Werkzeug 2.4. Use ``urllib.parse.parse_qs`` instead.
+        Will be removed in Werkzeug 3.0. Use ``urllib.parse.parse_qs`` instead.
 
     .. versionchanged:: 2.1
         The ``decode_keys`` parameter was removed.
@@ -1171,9 +1180,7 @@ def url_decode_stream(
 
         cls = MultiDict
 
-    with warnings.catch_warnings():
-        warnings.filterwarnings("ignore", "'make_chunk_iter", DeprecationWarning)
-        return cls(decoder)
+    return cls(decoder)
 
 
 def _url_decode_impl(